+++ app/states/static-states/player-halo-presentation-state.js
@@ -1,33 +1,18 @@
 let utils = require('../../helpers/utils');
-let { loadSoundPreferences } = require('../../helpers/audio');
 
 // TODO(piecioshka): może uda się nie robić specjalnego ekranu dla bohatera z halo (można wykorzystać PlayerPresentationState)
 
 export default class PlayerHaloPresentationState extends Phaser.State {
-    audio = {
-        scouter: null
-    };
-
     create() {
         this.add.image(0, 0, 'bg-player-presentation');
 
         this.cacheKey = `${this.game.player.id}-halo`;
 
-        this._setupSound();
         this._displayPlayer();
-
-        loadSoundPreferences(this.game);
-
-        // Dodajemy efekty audio.
-        this.audio.scouter.play();
 
         utils.timeout(this, Phaser.Timer.SECOND, () => {
             this.game.emit('game:player-halo-presents');
         });
-    }
-
-    _setupSound() {
-        this.audio.scouter = this.add.audio('scouter');
     }
 
     _displayPlayer() {