+++ app/states/static-states/shenron-state.js
@@ -1,11 +1,6 @@
 let utils = require('../../helpers/utils');
-let { loadSoundPreferences } = require('../../helpers/audio');
 
 export default class ShenronState extends Phaser.State {
-    audio = {
-        ambienceThunder: null
-    };
-
     create() {
         let background = this.add.image(0, 0, 'bg-shenron-growing');
 
@@ -13,18 +8,8 @@ export default class ShenronState extends Phaser.State {
             background.loadTexture('bg-shenron');
         });
 
-        this._setupSound();
-
-        loadSoundPreferences(this.game);
-
-        this.audio.ambienceThunder.play();
-
         utils.timeout(this, Phaser.Timer.SECOND * 3, () => {
             this.game.emit('game:shenron');
         });
-    }
-
-    _setupSound() {
-        this.audio.ambienceThunder = this.add.audio('sound-ambience-thunder');
     }
 }